requirements.txt
python-dotenv
uvicorn
fastapi
asyncio
openai
soundfile
websockets
# pip install 'uvicorn[standard]' fastapi openai edge-tts python-dotenv